removed big core

+++ cache/cache.h
@@ -2,6 +2,7 @@
 #define F_CACHE
 
 #include <stdbool.h>
+#include <sys/types.h>
 #ifdef MULTITHREADED
     #include <pthread.h>
 #endif
@@ -9,8 +10,8 @@
 typedef struct cache_entry{
     char* key;
     char* value;
-    int capacity;
-    int size;
+    size_t capacity;
+    size_t size;
     double resize_cf;
     bool finished;
 #ifdef MULTITHREADED
@@ -21,8 +22,8 @@ typedef struct cache_entry{
 
 typedef struct cache{
     cache_entry** entries;
-    int capacity;
-    int size;
+    size_t capacity;
+    size_t size;
 #ifdef MULTITHREADED
     pthread_mutex_t add_mutex;
     pthread_mutex_t size_mutex;
@@ -35,8 +36,8 @@ extern cache pr_cache;
 int init_cache();
 cache_entry* add_entry(char* key);
 cache_entry* find_entry_by_key(const char* const key);
-int append_entry(cache_entry* entry, char* data, int data_length);
-int send_entry_to_socket(cache_entry* entry, int socket, int progress);
+int append_entry(cache_entry* entry, char* data, size_t data_length);
+int send_entry_to_socket(cache_entry* entry, int socket, size_t progress);
 int remove_entry_by_key(const char* const key);
 int resize_entry(cache_entry* entry);
 int resize_cache();
